requirements.txt
pytest==7.2.2
selenium==4.8.3
sqlalchemy-stubs==0.4
webdriver-manager==3.8.6
requests==2.28.2
allure-pytest==2.13.1

pymongo~=4.3.3

SQLAlchemy~=2.0.9
PyMySQL==1.0.3
